requirements.txt
black==24.4.2
fire==0.6.0
google-generativeai == 0.5.4
pypdf == 4.2.0
pytest==8.2.0
pytest-cov==5.0.0
pylint==3.2.0
python-dotenv == 1.0.1
streamlit == 1.34.0
textblob == 0.18.0.post0